add option to create recipes with one text

+++ src/app/services/photo.service.ts
@@ -16,13 +16,10 @@ export class PhotoService {
   upload(photo: File): Observable<UploadResult> {
     // Remove spaces from name
     const photoName = photo.name.replace(/\s/g, '');
-    
+
     const result = new UploadResult();
     result.fileRef = this.storage.ref(`${this.FILE_PATH}/${photoName}`);
-    result.uploadTask = this.storage.upload(
-      `${this.FILE_PATH}/${photoName}`,
-      photo
-    );
+    result.uploadTask = this.storage.upload(`${this.FILE_PATH}/${photoName}`, photo);
     result.uploadPercentage$ = result.uploadTask.percentageChanges();
 
     // Get the downloadURL when the upload is finilized.
@@ -36,10 +33,9 @@ export class PhotoService {
   }
 
   delete(recipe: Recipe): Observable<void> {
-    const fileName = recipe.photoURL.substring(
-      recipe.photoURL.lastIndexOf('%2F') + 3,
-      recipe.photoURL.indexOf('?')
-    );
-    return this.storage.ref(`photos/${fileName}`).delete();
+    const fileName = recipe.photoURL?.substring(recipe.photoURL.lastIndexOf('%2F') + 3, recipe.photoURL.indexOf('?'));
+    if (fileName) {
+      return this.storage.ref(`photos/${fileName}`).delete();
+    }
   }
 }
